refactor: separate out ViewUserWindow
