+++ frontend/src/pages/roles/Roles.tsx
@@ -18,6 +18,8 @@ import TabManager from '../../utility_components/tabs';
 import { ViewPermissions } from './components/ViewPermissions';
 import { ViewRolesWindow } from './components/ViewRolesWindow';
 import { UserRoleRow } from './components/UserRoleRow';
+import { ViewUserWindow } from './components/ViewUserWindow';
+import { Wowzer } from './components/Wowzer';
 
 function Roles() {
     const userContext = useUserContext();
@@ -138,155 +140,7 @@ function Roles() {
                                             return <div key={index}>{val}</div>;
                                         });
                                     } else {
-                                        return (
-                                            <TabManager
-                                                orientation="row"
-                                                content={[{ tabTitle: '1' }, { tabTitle: '2' }]}
-                                                loader={() => {
-                                                    return (
-                                                        <TabManager
-                                                            orientation="column"
-                                                            content={[{ tabTitle: '1' }, { tabTitle: '2' }]}
-                                                            loader={() => {
-                                                                return (
-                                                                    <TabManager
-                                                                        orientation="row"
-                                                                        content={[{ tabTitle: '1' }, { tabTitle: '2' }]}
-                                                                        loader={(data) => {
-                                                                            return (
-                                                                                <TabManager
-                                                                                    orientation="column"
-                                                                                    content={[
-                                                                                        { tabTitle: '1' },
-                                                                                        { tabTitle: '2' },
-                                                                                    ]}
-                                                                                    loader={() => {
-                                                                                        return (
-                                                                                            <TabManager
-                                                                                                orientation="row"
-                                                                                                content={[
-                                                                                                    { tabTitle: '1' },
-                                                                                                    { tabTitle: '2' },
-                                                                                                ]}
-                                                                                                loader={(data) => {
-                                                                                                    return (
-                                                                                                        <TabManager
-                                                                                                            orientation="column"
-                                                                                                            content={[
-                                                                                                                {
-                                                                                                                    tabTitle:
-                                                                                                                        '1',
-                                                                                                                },
-                                                                                                                {
-                                                                                                                    tabTitle:
-                                                                                                                        '2',
-                                                                                                                },
-                                                                                                            ]}
-                                                                                                            loader={() => {
-                                                                                                                return (
-                                                                                                                    <TabManager
-                                                                                                                        orientation="row"
-                                                                                                                        content={[
-                                                                                                                            {
-                                                                                                                                tabTitle:
-                                                                                                                                    '1',
-                                                                                                                            },
-                                                                                                                            {
-                                                                                                                                tabTitle:
-                                                                                                                                    '2',
-                                                                                                                            },
-                                                                                                                        ]}
-                                                                                                                        loader={(
-                                                                                                                            data,
-                                                                                                                        ) => {
-                                                                                                                            return (
-                                                                                                                                <TabManager
-                                                                                                                                    orientation="column"
-                                                                                                                                    content={[
-                                                                                                                                        {
-                                                                                                                                            tabTitle:
-                                                                                                                                                '1',
-                                                                                                                                        },
-                                                                                                                                        {
-                                                                                                                                            tabTitle:
-                                                                                                                                                '2',
-                                                                                                                                        },
-                                                                                                                                    ]}
-                                                                                                                                    loader={() => {
-                                                                                                                                        return (
-                                                                                                                                            <TabManager
-                                                                                                                                                orientation="row"
-                                                                                                                                                content={[
-                                                                                                                                                    {
-                                                                                                                                                        tabTitle:
-                                                                                                                                                            '1',
-                                                                                                                                                    },
-                                                                                                                                                    {
-                                                                                                                                                        tabTitle:
-                                                                                                                                                            '2',
-                                                                                                                                                    },
-                                                                                                                                                    {
-                                                                                                                                                        tabTitle:
-                                                                                                                                                            '3',
-                                                                                                                                                        tabData:
-                                                                                                                                                            {
-                                                                                                                                                                content:
-                                                                                                                                                                    'Hello',
-                                                                                                                                                            },
-                                                                                                                                                    },
-                                                                                                                                                ]}
-                                                                                                                                                loader={(
-                                                                                                                                                    data,
-                                                                                                                                                ) => {
-                                                                                                                                                    if (
-                                                                                                                                                        data.tabTitle !=
-                                                                                                                                                        '3'
-                                                                                                                                                    ) {
-                                                                                                                                                        return (
-                                                                                                                                                            <>
-                                                                                                                                                                {
-                                                                                                                                                                    data.tabTitle
-                                                                                                                                                                }
-                                                                                                                                                            </>
-                                                                                                                                                        );
-                                                                                                                                                    } else {
-                                                                                                                                                        return (
-                                                                                                                                                            <>
-                                                                                                                                                                {
-                                                                                                                                                                    data
-                                                                                                                                                                        .tabData
-                                                                                                                                                                        .content
-                                                                                                                                                                }
-                                                                                                                                                            </>
-                                                                                                                                                        );
-                                                                                                                                                    }
-                                                                                                                                                }}
-                                                                                                                                            />
-                                                                                                                                        );
-                                                                                                                                    }}
-                                                                                                                                />
-                                                                                                                            );
-                                                                                                                        }}
-                                                                                                                    />
-                                                                                                                );
-                                                                                                            }}
-                                                                                                        />
-                                                                                                    );
-                                                                                                }}
-                                                                                            />
-                                                                                        );
-                                                                                    }}
-                                                                                />
-                                                                            );
-                                                                        }}
-                                                                    />
-                                                                );
-                                                            }}
-                                                        />
-                                                    );
-                                                }}
-                                            />
-                                        );
+                                        return <Wowzer />;
                                     }
                                 }}
                             />
@@ -307,71 +161,3 @@ function Roles() {
 }
 
 export default Roles;
-
-function ViewUserWindow({
-    users,
-    setUsers,
-    roles,
-}: {
-    users: Record<string, UserDTO>;
-    setUsers: (users: Record<string, UserDTO>) => void;
-    roles: Record<string, RoleDTO>;
-}) {
-    const content = Object.keys(users).map((username) => {
-        return {
-            tabTitle: users[username].name,
-            tabData: { ...users[username] },
-        };
-    });
-
-    const userContext = useUserContext();
-    async function removeRole(role: RoleDTO, userId: string) {
-        if (['Admin', 'Default'].indexOf(role.name) !== -1) {
-            return;
-        }
-
-        const data = await userContext.fetcher(
-            'DELETE /api/users/:userId/roles',
-            {
-                roleIds: [role.id],
-            },
-            { userId },
-        );
-
-        if (data) {
-            const newUsers = { ...users, [userId]: data.user };
-            setUsers(newUsers);
-        }
-    }
-
-    async function addRole(role: RoleDTO, userId: string) {
-        const data = await userContext.fetcher(
-            'POST /api/users/:userId/roles',
-            {
-                roleIds: [role.id],
-            },
-            { userId },
-        );
-
-        if (data) {
-            const newUsers = { ...users, [userId]: data.user };
-            setUsers(newUsers);
-        }
-    }
-
-    return (
-        <div className="p-4 flex flex-col gap-3 overflow-scroll w-full">
-            {content.map((data, index) => {
-                return (
-                    <UserRoleRow
-                        key={index}
-                        roles={Object.values(roles)}
-                        user={data.tabData}
-                        removeRole={removeRole}
-                        addRole={addRole}
-                    />
-                );
-            })}
-        </div>
-    );
-}
